Refactor item's copy number to a single field
